+++ lib/api/github.js
@@ -124,7 +124,7 @@ async function getRepos(token, endpoint) {
 
 // Initialize GitHub by getting base branch and SHA
 async function initRepo(repoName, token, endpoint) {
-  logger.debug(`initRepo(${repoName})`);
+  logger.debug(`initRepo(${JSON.stringify(repoName)})`);
   if (token) {
     process.env.GITHUB_TOKEN = token;
   } else if (!process.env.GITHUB_TOKEN) {
