+++ lib/config/index.js
@@ -10,12 +10,8 @@ const envParser = require('./env');
 
 const githubAppHelper = require('../helpers/github-app');
 
-let config = null;
-
 module.exports = {
   parseConfigs,
-  getCascadedConfig,
-  getRepositories,
 };
 
 async function parseConfigs(env, argv) {
@@ -27,7 +23,13 @@ async function parseConfigs(env, argv) {
   const cliConfig = cliParser.getConfig(argv);
   const envConfig = envParser.getConfig(env);
 
-  config = Object.assign({}, defaultConfig, fileConfig, envConfig, cliConfig);
+  const config = Object.assign(
+    {},
+    defaultConfig,
+    fileConfig,
+    envConfig,
+    cliConfig
+  );
 
   // Set log level
   logger.levels('stdout', config.logLevel);
@@ -74,8 +76,7 @@ async function parseConfigs(env, argv) {
     }
     config.repositories = await githubAppHelper.getRepositories(config);
     logger.info(`Found ${config.repositories.length} repositories installed`);
-    delete config.githubAppKey;
-    logger.debug(`GitHub App config: ${JSON.stringify(config)}`);
+    // TODO: redact logger.debug(`GitHub App config: ${JSON.stringify(config)}`);
   } else if (config.autodiscover) {
     // Autodiscover list of repositories
     if (config.platform === 'github') {
@@ -96,7 +97,7 @@ async function parseConfigs(env, argv) {
       logger.info(
         'The account associated with your token does not have access to any repos'
       );
-      return;
+      return config;
     }
   } else if (!config.repositories || config.repositories.length === 0) {
     // We need at least one repository defined
@@ -105,51 +106,12 @@ async function parseConfigs(env, argv) {
     );
   }
 
-  // Configure each repository
-  config.repositories = config.repositories.map(item => {
-    // Convert any repository strings to objects
-    const repo = typeof item === 'string' ? { repository: item } : item;
-
-    // copy across some fields from the base config if not present
-    repo.token = repo.token || config.token;
-    repo.platform = repo.platform || config.platform;
-    repo.onboarding = repo.onboarding || config.onboarding;
-    repo.endpoint = repo.endpoint || config.endpoint;
-
-    // Set default packageFiles
-    if (!repo.packageFiles || !repo.packageFiles.length) {
-      repo.packageFiles = config.packageFiles;
-    }
-
-    // Expand packageFile format
-    repo.packageFiles = repo.packageFiles.map(packageFile => {
-      if (typeof packageFile === 'string') {
-        return { fileName: packageFile };
-      }
-      return packageFile;
-    });
-
-    return repo;
-  });
-
   // Print config
   logger.debug(`config=${redact(config)}`);
   // Remove log file entries
   delete config.logFile;
   delete config.logFileLevel;
-}
-
-function getCascadedConfig(repo, packageFile) {
-  const cascadedConfig = Object.assign({}, config, repo, packageFile);
-  // Remove unnecessary fields
-  delete cascadedConfig.repositories;
-  delete cascadedConfig.repository;
-  delete cascadedConfig.fileName;
-  return cascadedConfig;
-}
-
-function getRepositories() {
-  return config.repositories;
+  return config;
 }
 
 function redact(inputConfig) {
